+++ linux/services/mqtt-subscriptions/commandCallback.cpp
@@ -40,8 +40,9 @@ void CommandCallback::publishMessage(std::string topic, std::string message)
     try {
         mqttTopic.publish(std::move(message));
     }
-    catch (const mqtt::exception& exc) {
-        std::cerr << exc.what() << std::endl;
+    catch (std::exception const& e) {
+        std::cout << "ERROR: publishMessage" << std::endl;
+        std::cerr << e.what() << std::endl;
         std::cerr << "topic: " << topic << std::endl;
         std::cerr << "message: " << message << std::endl;
     }
@@ -67,8 +68,6 @@ std::vector<std::string> CommandCallback::splitString(std::string s, const std::
 
 void CommandCallback::pollNode(std::string commandName, uint32_t nodeAddress)
 {
-    // std::cout << "DEBUG: pollNode(" << commandName << ", " << std::to_string(nodeAddress) << ")" << std::endl;
-
     std::string topic = "radio-arduino/RCMD/proxy/" + std::to_string(nodeAddress);
     std::string message = "{\"name\": \"" + commandName + "\", \"nodeAddress\": " + std::to_string(nodeAddress)
         + ", \"expirationTime\": 0}";
@@ -80,50 +79,47 @@ void CommandCallback::executeSubscriptions()
 {
     std::vector<uint32_t> radioNodes = getRadioNodeIdList();
 
-    // std::cout << "DEBUG: executeSubscriptions" << std::endl;
-    for (int i = 0; i < m_nodeTwins.size(); i++) {
-        // std::cout << "DEBUG: execute subscriptions for node " << std::to_string(nodeTwin.nodeAddress) << std::endl;
-
-        for (auto subscription : m_subscriptions) {
+    for (int i = 0; i < m_nodeInfoList.size(); i++) {
+        bool nodeAppearToBeStuck = false;
+        for (int j = 0; j < m_subscriptions.size(); j++) {
             int loopCounter = 0;
-            while (false == m_nodeTwins.at(i).readyForCommand) {
-                if (loopCounter > 1000) {
+            while (false == m_nodeInfoList.at(i).readyForCommand) {
+                if (loopCounter++ > 1000) {
+                    nodeAppearToBeStuck = true;
                     break;
                 }
-                loopCounter++;
                 std::this_thread::sleep_for(10ms);
-                // std::cout << "DEBUG: wait loop" << std::endl;
             }
 
-            if (subscription.nodeAddress == m_nodeTwins.at(i).nodeAddress) {
-                if (m_nodeTwins.at(i).readyForCommand) {
+            if (m_subscriptions.at(j).nodeAddress == m_nodeInfoList.at(i).nodeAddress) {
+                if (m_nodeInfoList.at(i).readyForCommand) {
                     using namespace std::chrono;
                     auto start = std::chrono::high_resolution_clock::now();
                     uint64_t time_since_epoch_ms
                         = std::chrono::duration_cast<std::chrono::milliseconds>(start.time_since_epoch()).count();
 
-                    int diff = time_since_epoch_ms - subscription.timestampLastValidResponse;
-                    // std::cout << "diff: " << std::to_string(diff) << std::endl;
+                    int diff = time_since_epoch_ms - m_subscriptions.at(j).timestampLastValidResponse;
 
-                    if ((diff > subscription.intervalInSeconds * 1000) || diff < 0) {
-                        pollNode(subscription.commandName, subscription.nodeAddress);
-                        m_nodeTwins.at(i).readyForCommand = false;
-                        std::this_thread::sleep_for(10000ms); // TODO: do we need this?
+                    if ((diff > m_subscriptions.at(j).intervalInSeconds * 1000) || diff < 0) {
+                        pollNode(m_subscriptions.at(j).commandName, m_subscriptions.at(j).nodeAddress);
+                        m_nodeInfoList.at(i).readyForCommand = false;
                     }
                 }
-                else {
-                    std::cout << "WARNING: node '" << std::to_string(subscription.nodeAddress)
-                              << "' not ready for command" << std::endl;
-                }
+                // else {
+                // std::cout << "WARNING: node '" << std::to_string(m_subscriptions.at(j).nodeAddress)
+                //<< "' not ready for command" << std::endl;
+                // break;
+                //}
             }
         }
-        if (false == m_nodeTwins.at(i).readyForCommand) {
-            m_nodeTwins.at(i).readyForCommand = true;
-            std::this_thread::sleep_for(1000ms);
+        if (nodeAppearToBeStuck) {
+            std::this_thread::sleep_for(15000ms); // wait from response before switching to the next node
+            std::cout << "DEBUG: force readyForCommand=true, address="
+                      << std::to_string(m_nodeInfoList.at(i).nodeAddress) << std::endl;
+            m_nodeInfoList.at(i).readyForCommand = true;
         }
+            std::this_thread::sleep_for(15000ms); // wait from response before switching to the next node
     }
-
-    std::this_thread::sleep_for(10000ms);
 }
 
 void CommandCallback::message_arrived(mqtt::const_message_ptr message)
@@ -138,8 +134,6 @@ void CommandCallback::message_arrived(mqtt::const_message_ptr message)
         try {
             auto jsonData = json::parse(payload);
             uint32_t nodeAddress = jsonData["nodeAddress"];
-            int healthIndicator = jsonData["healthIndicator"].get<int>();
-            uint64_t lastAdvertisement = jsonData["lastAdvertisement"].get<uint64_t>();
 
             bool nodeExists
                 = std::find(m_radioNodeIdList.begin(), m_radioNodeIdList.end(), nodeAddress) != m_radioNodeIdList.end();
@@ -147,31 +141,26 @@ void CommandCallback::message_arrived(mqtt::const_message_ptr message)
             if (!nodeExists) {
                 std::cout << "Added radio node: " << std::to_string(nodeAddress) << std::endl;
                 m_radioNodeIdList.push_back(nodeAddress);
-                nodeTwin_t nodeTwin;
-                nodeTwin.nodeAddress = nodeAddress;
-                nodeTwin.readyForCommand = true;
-                m_nodeTwins.push_back(nodeTwin);
+                nodeInfo_t nodeInfo;
+                nodeInfo.nodeAddress = nodeAddress;
+                nodeInfo.readyForCommand = true;
+                m_nodeInfoList.push_back(nodeInfo);
                 // add subscriptions
                 if (topic_orig.starts_with(
                         "radio-arduino/DBIRTH/" + std::to_string(gatewayAddress) + "/"
                         + std::to_string(gatewayAddress))) {
-                    // std::cout << "DEBUG: add subscription for gatway" << std::endl;
-
-                    m_subscriptions.push_back({ "get_device_name", 24 * 60 * 60, nodeAddress });
-                    //m_subscriptions.push_back({ "get_attached_devices_csv_string", 24 * 60 * 60, nodeAddress });
+                    m_subscriptions.push_back({ "get_device_name", 60 * 120, nodeAddress });
+                    m_subscriptions.push_back({ "get_attached_devices_csv_string", 60 * 120, nodeAddress });
                     m_subscriptions.push_back({ "get_statistics", 60 * 120, nodeAddress });
-                    m_subscriptions.push_back({ "get_version", 24 * 60 * 60, nodeAddress });
+                    m_subscriptions.push_back({ "get_version", 60 * 120, nodeAddress });
                 }
                 else {
-                    // std::cout << "DEBUG: add subscription for node" << std::endl;
-
                     m_subscriptions.push_back({ "vcc", 60 * 10, nodeAddress });
                     m_subscriptions.push_back({ "gpio", 60 * 60, nodeAddress });
                     m_subscriptions.push_back({ "get_active_time_counter", 60, nodeAddress });
-                    m_subscriptions.push_back({ "get_device_name", 24 * 60 * 60, nodeAddress });
+                    m_subscriptions.push_back({ "get_device_name", 60 * 120, nodeAddress });
                     m_subscriptions.push_back({ "get_statistics", 60 * 120, nodeAddress });
-                    m_subscriptions.push_back({ "get_attached_devices_csv_string", 24 * 60 * 60, nodeAddress });
-                    // m_subscriptions.push_back({ "get_version", 24*60*60, nodeAddress });
+                    m_subscriptions.push_back({ "get_attached_devices_csv_string", 60 * 120, nodeAddress });
                 }
             }
         }
@@ -182,18 +171,16 @@ void CommandCallback::message_arrived(mqtt::const_message_ptr message)
     }
 
     if (topic_orig.starts_with("radio-arduino/STATE/")) {
-        auto tokens = splitString(topic_orig, "/");
         std::string nodeAddressString = tokens.at(3);
+        uint32_t nodeAddress = stoul(tokens.at(3));
 
-        for (int i = 0; i < m_nodeTwins.size(); i++) {
-            if ((std::to_string(m_subscriptions.at(i).nodeAddress) == nodeAddressString)) {
+        for (int i = 0; i < m_nodeInfoList.size(); i++) {
+            if (m_nodeInfoList.at(i).nodeAddress == nodeAddress) {
                 if (payload.empty()) {
-                    m_nodeTwins.at(i).readyForCommand = true;
-                    // std::cout << "DEBUG: readyForCommand(" << nodeAddressString << ")=true" << std::endl;
+                    m_nodeInfoList.at(i).readyForCommand = true;
                 }
                 else {
-                    m_nodeTwins.at(i).readyForCommand = false;
-                    // std::cout << "DEBUG: readyForCommand(" << nodeAddressString << ")=false" << std::endl;
+                    m_nodeInfoList.at(i).readyForCommand = false;
                 }
             }
         }
@@ -201,41 +188,9 @@ void CommandCallback::message_arrived(mqtt::const_message_ptr message)
 
     if (topic_orig.starts_with("radio-arduino/DDATA/")) {
         try {
-            auto tokens = splitString(topic_orig, "/");
-            if(tokens.size() < 5)
-            {
-                std::cout << "DEBUG: ABORT! parsing" << std::endl;
-            }
-            uint32_t nodeAddress = std::stoul(tokens.at(3));
-            std::string command_name = tokens.at(4);
-
-            if (command_name == "get_attached_devices_csv_string") {
-                auto jsonData = json::parse(payload);
-                std::string csvString = jsonData["payload"]["csvString"];
-                if (csvString.find("lsm303d") != std::string::npos) {
-                    m_subscriptions.push_back({ "get_lsm303d", 60, nodeAddress });
-                }
-                else if (csvString.find("quad") != std::string::npos) {
-                    m_subscriptions.push_back({ "quadrature_encoder", 60, nodeAddress });
-                }
-                else if (csvString.find("ina219") != std::string::npos) {
-                    m_subscriptions.push_back({ "ina219", 60, nodeAddress });
-                }
-            }
-
-            // seher
-        }
-        catch (std::exception const& e) {
-            std::cout << "DEBUG: malformed DDATA" << std::endl;
-        }
-
-        try {
             auto jsonData = json::parse(payload);
 
-            auto tokens = splitString(topic_orig, "/");
             std::string nodeAddressString = tokens.at(3);
-            // std::cout << "DEBUG: nodeAddressString=" << nodeAddressString << std::endl;
-
             std::string responseCode = jsonData["responseCode"];
             std::string commandName = jsonData["name"];
             uint64_t timestamp = jsonData["timestamp"].get<uint64_t>();
@@ -245,22 +200,40 @@ void CommandCallback::message_arrived(mqtt::const_message_ptr message)
                     && (responseCode == "success") && (commandName == m_subscriptions.at(i).commandName)
 
                 ) {
+                    if (commandName == "get_attached_devices_csv_string") {
+                        uint32_t nodeAddress = std::stoul(nodeAddressString);
+                        auto jsonData = json::parse(payload);
+                        std::string csvString = jsonData["payload"]["csvString"];
+                        if (csvString.find("lsm303d") != std::string::npos) {
+                            // TODO: make method that check if exist before adding
+                            m_subscriptions.push_back({ "get_lsm303d", 60*5, nodeAddress });
+                        }
+                        else if (csvString.find("quad") != std::string::npos) {
+                            m_subscriptions.push_back({ "quadrature_encoder", 60*5, nodeAddress });
+                        }
+                        else if (csvString.find("ina219") != std::string::npos) {
+                            m_subscriptions.push_back({ "ina219", 5, nodeAddress });
+                        }
+                    }
+
                     m_subscriptions.at(i).lastValidResponse = payload;
                     m_subscriptions.at(i).timestampLastValidResponse = timestamp;
-                    // std::cout << "DEBUG: -----> saveResponse(" << commandName << ", " << nodeAddressString << ")"
-                    //<< std::endl;
+
+                    std::string republish_topic = "raduino-subscriptions/DDATA/"
+                        + std::to_string(m_subscriptions.at(i).nodeAddress) + "/" + commandName;
+
+                    publishMessage(republish_topic, payload);
                 }
             }
 
-            for (int i = 0; i < m_nodeTwins.size(); i++) {
-                if (nodeAddressString == std::to_string(m_nodeTwins.at(i).nodeAddress)) {
-                    // we got respnse so node is ready for new command
-                    m_nodeTwins.at(i).readyForCommand = true;
+            for (int i = 0; i < m_nodeInfoList.size(); i++) {
+                if (nodeAddressString == std::to_string(m_nodeInfoList.at(i).nodeAddress)) {
+                    m_nodeInfoList.at(i).readyForCommand = true; // got response from node
                 }
             }
         }
         catch (std::exception const& e) {
-            std::cout << "DEBUG: malformed DDATA" << std::endl;
+            std::cout << "DEBUG: malformed DDATA. " << e.what() << std::endl;
         }
     }
 }